DuAn/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
blinker==1.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.1.3; python_version >= '3.7'
flask-cors==3.0.10
flask-mail==0.9.1
flask-marshmallow==0.14.0
flask-pymongo==2.3.0
flask==2.1.3
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
marshmallow==3.17.0
mongoalchemy==0.21
packaging==21.3; python_version >= '3.6'
pyjwt==2.4.0
pymongo==4.2.0; python_version >= '3.7'
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dotenv==0.20.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.39
werkzeug==2.1.2; python_version >= '3.7'
